@@ -1,6 +1,6 @@
 import { defineClientConfig } from '@vuepress/client'
 import { VueLive } from "vue-live";
-import "vue-live/lib/vue-live.esm.css";
+import './vue-prism-editor.prismeditor.min.css';
 
 export default defineClientConfig({
   enhance({ app }) {
